Reuse DoReqAndVerifyStatusCode
